Extract the node classes into their own files
