+++ Tests/TuistGeneratorTests/Graph/GraphNodeTests.swift
@@ -8,8 +8,8 @@ import TuistCoreTesting
 final class GraphNodeTests: XCTestCase {
     func test_set() {
         // Given
-        let a = GraphNode(path: "/path/a")
-        let b = GraphNode(path: "/path/b")
+        let a = GraphNode(path: "/path/a", name: "a")
+        let b = GraphNode(path: "/path/b", name: "b")
         let c1 = TargetNode(project: .test(path: "/path/c"),
                             target: .test(name: "c1"),
                             dependencies: [])
@@ -34,9 +34,9 @@ final class GraphNodeTests: XCTestCase {
 
     func test_equality() {
         // Given
-        let a1 = GraphNode(path: "/a")
-        let a2 = GraphNode(path: "/a")
-        let b = GraphNode(path: "/b")
+        let a1 = GraphNode(path: "/a", name: "a")
+        let a2 = GraphNode(path: "/a", name: "a")
+        let b = GraphNode(path: "/b", name: "b")
 
         // When / Then
         XCTAssertEqual(a1, a2)
@@ -46,7 +46,7 @@ final class GraphNodeTests: XCTestCase {
 
     func test_subclass_equality() {
         // Given
-        let a = GraphNode(path: "/a")
+        let a = GraphNode(path: "/a", name: "a")
         let b = TargetNode(project: .test(path: "/a"), target: .test(), dependencies: [])
         let c = LibraryNode(path: "/a", publicHeaders: "/path/to/headers")
 
@@ -61,281 +61,5 @@ final class GraphNodeTests: XCTestCase {
                 XCTAssertNotEqual(lhs.element, rhs.element)
             }
         }
-    }
-}
-
-final class TargetNodeTests: XCTestCase {
-    func test_equality() {
-        // Given
-        let c1 = TargetNode(project: .test(path: "/c"),
-                            target: .test(name: "c"),
-                            dependencies: [])
-        let c2 = TargetNode(project: .test(path: "/c"),
-                            target: .test(name: "c"),
-                            dependencies: [])
-        let c3 = TargetNode(project: .test(path: "/c"),
-                            target: .test(name: "c3"),
-                            dependencies: [])
-        let d = TargetNode(project: .test(path: "/d"),
-                           target: .test(name: "c"),
-                           dependencies: [])
-
-        // When / Then
-        XCTAssertEqual(c1, c2)
-        XCTAssertNotEqual(c2, c3)
-        XCTAssertNotEqual(c1, d)
-        XCTAssertEqual(d, d)
-    }
-
-    func test_equality_asGraphNodes() {
-        // Given
-        let c1: GraphNode = TargetNode(project: .test(path: "/c"),
-                                       target: .test(name: "c"),
-                                       dependencies: [])
-        let c2: GraphNode = TargetNode(project: .test(path: "/c"),
-                                       target: .test(name: "c"),
-                                       dependencies: [])
-        let c3: GraphNode = TargetNode(project: .test(path: "/c"),
-                                       target: .test(name: "c3"),
-                                       dependencies: [])
-        let d: GraphNode = TargetNode(project: .test(path: "/d"),
-                                      target: .test(name: "c"),
-                                      dependencies: [])
-
-        // When / Then
-        XCTAssertEqual(c1, c2)
-        XCTAssertNotEqual(c2, c3)
-        XCTAssertNotEqual(c1, d)
-    }
-
-    func test_encode() {
-        // Given
-        let library = LibraryNode.test()
-        let framework = FrameworkNode.test()
-        let node = TargetNode(project: .test(path: "/"),
-                              target: .test(name: "Target"),
-                              dependencies: [library, framework])
-
-        let expected = """
-        {
-            "type": "source",
-            "path" : "\(node.path.pathString)",
-            "bundle_id" : "\(node.target.bundleId)",
-            "product" : "\(node.target.product.rawValue)",
-            "name" : "\(node.target.name)",
-            "dependencies" : [
-                "\(library.name)",
-                "\(framework.name)"
-            ],
-            "platform" : "\(node.target.platform.rawValue)"
-        }
-        """
-
-        // Then
-        XCTAssertEncodableEqualToJson(node, expected)
-    }
-}
-
-final class PrecompiledNodeTests: XCTestCase {
-    var system: MockSystem!
-
-    override func setUp() {
-        super.setUp()
-        system = MockSystem()
-    }
-
-    func test_architecture_rawValues() {
-        XCTAssertEqual(PrecompiledNode.Architecture.x8664.rawValue, "x86_64")
-        XCTAssertEqual(PrecompiledNode.Architecture.i386.rawValue, "i386")
-        XCTAssertEqual(PrecompiledNode.Architecture.armv7.rawValue, "armv7")
-        XCTAssertEqual(PrecompiledNode.Architecture.armv7s.rawValue, "armv7s")
-    }
-}
-
-final class FrameworkNodeTests: XCTestCase {
-    var system: MockSystem!
-    var subject: FrameworkNode!
-    var path: AbsolutePath!
-
-    override func setUp() {
-        super.setUp()
-        system = MockSystem()
-        path = AbsolutePath("/test.framework")
-        subject = FrameworkNode(path: path)
-    }
-
-    func test_name() {
-        XCTAssertEqual(subject.name, "test")
-    }
-
-    func test_binaryPath() {
-        XCTAssertEqual(subject.binaryPath.pathString, "/test.framework/test")
-    }
-
-    func test_isCarthage() {
-        XCTAssertFalse(subject.isCarthage)
-        subject = FrameworkNode(path: AbsolutePath("/path/Carthage/Build/iOS/A.framework"))
-        XCTAssertTrue(subject.isCarthage)
-    }
-
-    func test_architectures_when_nonFatFramework() throws {
-        system.succeedCommand("/usr/bin/lipo -info /test.framework/test",
-                              output: "Non-fat file: path is architecture: x86_64")
-        try XCTAssertEqual(subject.architectures(system: system).first, .x8664)
-    }
-
-    func test_architectures_when_fatFramework() throws {
-        system.succeedCommand("/usr/bin/lipo -info /test.framework/test",
-                              output: "Architectures in the fat file: /path/xpm.framework/xpm are: x86_64 arm64")
-        try XCTAssertTrue(subject.architectures(system: system).contains(.x8664))
-        try XCTAssertTrue(subject.architectures(system: system).contains(.arm64))
-    }
-
-    func test_linking() {
-        system.succeedCommand("/usr/bin/file", "/test.framework/test",
-                              output: "whatever dynamically linked")
-        try XCTAssertEqual(subject.linking(system: system), .dynamic)
-    }
-
-    func test_encode() {
-        // Given
-        let framework = FrameworkNode(path: fixturePath(path: RelativePath("xpm.framework")))
-        let expected = """
-        {
-            "path": "\(framework.path)",
-            "architectures": ["x86_64", "arm64"],
-            "name": "xpm",
-            "type": "precompiled",
-            "product": "framework"
-        }
-        """
-
-        // Then
-        XCTAssertEncodableEqualToJson(framework, expected)
-    }
-}
-
-final class LibraryNodeTests: XCTestCase {
-    var system: MockSystem!
-    var subject: LibraryNode!
-    var path: AbsolutePath!
-
-    override func setUp() {
-        super.setUp()
-        system = MockSystem()
-        path = AbsolutePath("/test.a")
-        subject = LibraryNode(path: path, publicHeaders: AbsolutePath("/headers"))
-    }
-
-    func test_name() {
-        XCTAssertEqual(subject.name, "test")
-    }
-
-    func test_binaryPath() {
-        XCTAssertEqual(subject.binaryPath.pathString, "/test.a")
-    }
-
-    func test_architectures() throws {
-        system.succeedCommand("/usr/bin/lipo", "-info", "/test.a", output: "Non-fat file: path is architecture: x86_64")
-        try XCTAssertEqual(subject.architectures(system: system).first, .x8664)
-    }
-
-    func test_linking() {
-        system.succeedCommand("/usr/bin/file", "/test.a", output: "whatever dynamically linked")
-        try XCTAssertEqual(subject.linking(system: system), .dynamic)
-    }
-
-    func test_equality() {
-        // Given
-        let a1 = LibraryNode(path: "/a", publicHeaders: "/a/header", swiftModuleMap: "/a/swiftmodulemap")
-        let a2 = LibraryNode(path: "/a", publicHeaders: "/a/header/2", swiftModuleMap: "/a/swiftmodulemap")
-        let b = LibraryNode(path: "/b", publicHeaders: "/b/header", swiftModuleMap: "/b/swiftmodulemap")
-
-        // When / Then
-        XCTAssertEqual(a1, a1)
-        XCTAssertNotEqual(a1, a2)
-        XCTAssertNotEqual(a2, b)
-        XCTAssertNotEqual(a1, b)
-    }
-
-    func test_encode() {
-        // Given
-        let library = LibraryNode(path: fixturePath(path: RelativePath("libStaticLibrary.a")),
-                                  publicHeaders: fixturePath(path: RelativePath("")))
-        let expected = """
-        {
-            "type": "precompiled",
-            "path" : "\(library.path)",
-            "architectures" : [
-                "x86_64"
-            ],
-            "name" : "\(library.name)",
-            "product" : "static_library"
-        }
-        """
-
-        // Then
-        XCTAssertEncodableEqualToJson(library, expected)
-    }
-}
-
-final class SDKNodeTests: XCTestCase {
-    func test_sdk_supportedTypes() throws {
-        // Given
-        let libraries = [
-            "Foo.framework",
-            "libBar.tbd",
-        ]
-
-        // When / Then
-        XCTAssertNoThrow(try libraries.map { try SDKNode(name: $0, platform: .macOS, status: .required) })
-    }
-
-    func test_sdk_usupportedTypes() throws {
-        XCTAssertThrowsError(try SDKNode(name: "FooBar", platform: .tvOS, status: .required)) { error in
-            XCTAssertEqual(error as? SDKNode.Error, .unsupported(sdk: "FooBar"))
-        }
-    }
-
-    func test_sdk_errors() {
-        XCTAssertEqual(SDKNode.Error.unsupported(sdk: "Foo").type, .abort)
-    }
-
-    func test_sdk_framework_paths() throws {
-        // Given
-        let libraries: [(name: String, platform: Platform)] = [
-            ("Foo.framework", .iOS),
-            ("Bar.framework", .macOS),
-            ("FooBar.framework", .tvOS),
-        ]
-
-        // When
-        let nodes = try libraries.map { try SDKNode(name: $0.name, platform: $0.platform, status: .required) }
-
-        // Then
-        XCTAssertEqual(nodes.map(\.path), [
-            "/Platforms/iPhoneOS.platform/Developer/SDKs/iPhoneOS.sdk/System/Library/Frameworks/Foo.framework",
-            "/Platforms/MacOSX.platform/Developer/SDKs/MacOSX.sdk/System/Library/Frameworks/Bar.framework",
-            "/Platforms/AppleTVOS.platform/Developer/SDKs/AppleTVOS.sdk/System/Library/Frameworks/FooBar.framework",
-        ])
-    }
-
-    func test_sdk_library_paths() throws {
-        // Given
-        let libraries: [(name: String, platform: Platform)] = [
-            ("libFoo.tbd", .iOS),
-            ("libBar.tbd", .macOS),
-            ("libFooBar.tbd", .tvOS),
-        ]
-
-        // When
-        let nodes = try libraries.map { try SDKNode(name: $0.name, platform: $0.platform, status: .required) }
-
-        // Then
-        XCTAssertEqual(nodes.map(\.path), [
-            "/Platforms/iPhoneOS.platform/Developer/SDKs/iPhoneOS.sdk/usr/lib/libFoo.tbd",
-            "/Platforms/MacOSX.platform/Developer/SDKs/MacOSX.sdk/usr/lib/libBar.tbd",
-            "/Platforms/AppleTVOS.platform/Developer/SDKs/AppleTVOS.sdk/usr/lib/libFooBar.tbd",
-        ])
     }
 }
